--updated test syntax

+++ src/tests/AttributesManagersTest.cpp
@@ -765,9 +765,9 @@ void AttributesManagersTest::testLightAttrVals(
   CORRADE_COMPARE(lightAttr0->getColor(), Magnum::Vector3(0.6, 0.7, 0.8));
 
   CORRADE_COMPARE(lightAttr0->getIntensity(), -0.1);
-  CORRADE_COMPARE(lightAttr0->getType(),
+  CORRADE_COMPARE(static_cast<int>(lightAttr0->getType()),
                   static_cast<int>(esp::gfx::LightType::Directional));
-  CORRADE_COMPARE(lightAttr0->getPositionModel(),
+  CORRADE_COMPARE(static_cast<int>(lightAttr0->getPositionModel()),
                   static_cast<int>(esp::gfx::LightPositionModel::Camera));
   CORRADE_COMPARE(lightAttr0->getInnerConeAngle(), 0.25_radf);
   CORRADE_COMPARE(lightAttr0->getOuterConeAngle(), -1.57_radf);
@@ -780,9 +780,9 @@ void AttributesManagersTest::testLightAttrVals(
   CORRADE_COMPARE(lightAttr1->getColor(), Magnum::Vector3(0.5, 0.3, 0.1));
 
   CORRADE_COMPARE(lightAttr1->getIntensity(), -1.2);
-  CORRADE_COMPARE(lightAttr1->getType(),
+  CORRADE_COMPARE(static_cast<int>(lightAttr1->getType()),
                   static_cast<int>(esp::gfx::LightType::Point));
-  CORRADE_COMPARE(lightAttr1->getPositionModel(),
+  CORRADE_COMPARE(static_cast<int>(lightAttr1->getPositionModel()),
                   static_cast<int>(esp::gfx::LightPositionModel::Global));
   CORRADE_COMPARE(lightAttr1->getInnerConeAngle(), -0.75_radf);
   CORRADE_COMPARE(lightAttr1->getOuterConeAngle(), -1.7_radf);
@@ -1184,7 +1184,7 @@ void AttributesManagersTest::testStageAttrVals(
   // stage-specific attributes
   CORRADE_COMPARE(stageAttr->getGravity(), Magnum::Vector3(9, 8, 7));
   // make sure that is not default value "flat"
-  CORRADE_COMPARE(stageAttr->getShaderType(),
+  CORRADE_COMPARE(static_cast<int>(stageAttr->getShaderType()),
                   static_cast<int>(Attrs::ObjectInstanceShaderType::Material));
   CORRADE_COMPARE(stageAttr->getOrigin(), Magnum::Vector3(1, 2, 3));
   CORRADE_COMPARE(stageAttr->getSemanticAssetHandle(),
@@ -1293,7 +1293,7 @@ void AttributesManagersTest::testObjectAttrVals(
   CORRADE_COMPARE(objAttr->getSemanticId(), 7);
   // object-specific attributes
   CORRADE_COMPARE(objAttr->getMass(), 9);
-  CORRADE_COMPARE(objAttr->getShaderType(),
+  CORRADE_COMPARE(static_cast<int>(objAttr->getShaderType()),
                   static_cast<int>(Attrs::ObjectInstanceShaderType::Phong));
   CORRADE_VERIFY(objAttr->getBoundingBoxCollisions());
   CORRADE_VERIFY(objAttr->getJoinCollisionMeshes());
